Conflicts:
	NetBeansProjects/Iris/src/graphs/Grapher.java
	NetBeansProjects/Iris/src/graphs/MemoryManager.java
	NetBeansProjects/Iris/src/graphs/MemoryManagerConsole.java
	NetBeansProjects/Iris/src/graphs/MemoryNode.java
	NetBeansProjects/Iris/src/graphs/MemoryNodeUI.java
	NetBeansProjects/Iris/src/statistics/ContinuousDistributionEditor.java

+++ src/statistics_distributions/ChiSquared.java
@@ -41,6 +41,6 @@ public class ChiSquared extends ContinuousDistribution {
 
     @Override
     public double est_param_impl(int i, DataSet data) {
-        return data.getMean();
+        return (int)data.getMean();
     }
 }
